Move chat types to separate package
